async

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -9,11 +9,11 @@ Use the random module.
 """
 
 
-import asyncio
 import random
+import asyncio
 
 
-async def wait_random(max_delay: float = 10.0) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     """
     Function - wait_random
